Add fallback for undefined UserComment strings
darktable and potentially any program that uses exiv2 will set the type
for the exif UserComment tag to undefined for utf-8 encoded strings.

exiftool will happily decode such a string.

This patch adds a fallback that tries to decode this type of strings as
utf-8, otherwise return the empty string.